+++ egas/inc/filter.h
@@ -11,6 +11,6 @@
 #include <stdint.h>
 #include "em_cmu.h"
 
-void EGAS_SavGol_Filter(uint16_t* _data);
+void EGAS_SavGol_Filter(uint32_t* _data);
 
 #endif /* EGAS_INC_FILTER_H_ */
